Refactored to DeathDB and changed a permission.
