+++ src/main/java/com/nearvanilla/deathdb/events/OnPlayerDeath.kt
@@ -1,7 +1,7 @@
 /* Licensed under GNU General Public License v3.0 */
-package com.nearvanilla.deathmanager.events
+package com.nearvanilla.deathdb.events
 
-import com.nearvanilla.deathmanager.DeathManager
+import com.nearvanilla.deathdb.DeathDB
 import org.bukkit.event.EventHandler
 import org.bukkit.event.Listener
 import org.bukkit.event.entity.PlayerDeathEvent
@@ -12,9 +12,9 @@ class OnPlayerDeath : Listener {
     fun onPlayerDeath(event: PlayerDeathEvent) {
         val player = event.player
         try {
-            DeathManager.dbWrapper.addDeathRecord(player)
+            DeathDB.dbWrapper.addDeathRecord(player)
         } catch (e: Exception) { // Could be SQLException but execute can also throw a timeout exception.
-            DeathManager.pluginLogger.warning("Error occurred on player death.\n${e.message}")
+            DeathDB.pluginLogger.warning("Error occurred on player death.\n${e.message}")
         }
     }
 }
