+++ src/main/java/com/nearvanilla/deathdb/exceptions/DeathDBException.kt
@@ -1,0 +1,4 @@
+/* Licensed under GNU General Public License v3.0 */
+package com.nearvanilla.deathdb.exceptions
+
+class DeathDBException(message: String) : Exception(message)
