+++ src/main/java/com/nearvanilla/deathdb/libs/DatabaseWrapper.kt
@@ -1,27 +1,27 @@
 /* Licensed under GNU General Public License v3.0 */
-package com.nearvanilla.deathmanager.libs
+package com.nearvanilla.deathdb.libs
 
-import com.nearvanilla.deathmanager.exceptions.DeathManagerException
+import com.nearvanilla.deathdb.exceptions.DeathDBException
 import org.bukkit.entity.Player
 import java.sql.DriverManager
 import java.sql.ResultSet
 import java.time.Instant
 
 class DatabaseWrapper(dbPath: String) {
-    private val databaseConnection = DriverManager.getConnection("jdbc:sqlite:$dbPath") ?: throw DeathManagerException("Failed to establish connection to database.")
+    private val databaseConnection = DriverManager.getConnection("jdbc:sqlite:$dbPath") ?: throw DeathDBException("Failed to establish connection to database.")
 
     fun createDeathsTable() {
         try {
             val createTableStmt =
                 databaseConnection.prepareStatement("CREATE TABLE IF NOT EXISTS deaths(id INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT, uniqueId TEXT NOT NULL, timeOfDeath INTEGER NOT NULL, posX REAL NOT NULL, posY REAL NOT NULL, posZ REAL NOT NULL, worldName TEXT NOT NULL, serializedInventory TEXT NOT NULL)")
-                    ?: throw DeathManagerException("Failed to prepare the create table statement.")
+                    ?: throw DeathDBException("Failed to prepare the create table statement.")
             createTableStmt.execute()
         } catch (e: Exception) {
         }
     }
 
     fun addDeathRecord(playerWhoDied: Player) {
-        val addDeathStmt = databaseConnection.prepareStatement("INSERT INTO deaths VALUES(NULL, ?, ?, ?, ?, ?, ?, ?)") ?: throw DeathManagerException("Failed to prepare the add death statement.")
+        val addDeathStmt = databaseConnection.prepareStatement("INSERT INTO deaths VALUES(NULL, ?, ?, ?, ?, ?, ?, ?)") ?: throw DeathDBException("Failed to prepare the add death statement.")
         addDeathStmt.setString(1, playerWhoDied.uniqueId.toString())
         addDeathStmt.setLong(2, Instant.now().epochSecond)
         addDeathStmt.setDouble(3, playerWhoDied.location.x)
@@ -33,7 +33,7 @@ class DatabaseWrapper(dbPath: String) {
     }
 
     fun getPlayerInformation(player: Player): ResultSet {
-        val getPlayerInfoStmt = databaseConnection.prepareStatement("SELECT * FROM deaths WHERE uniqueId = ? ORDER BY timeOfDeath DESC LIMIT 5") ?: throw DeathManagerException("Failed to prepare the get player info statement.")
+        val getPlayerInfoStmt = databaseConnection.prepareStatement("SELECT * FROM deaths WHERE uniqueId = ? ORDER BY timeOfDeath DESC LIMIT 5") ?: throw DeathDBException("Failed to prepare the get player info statement.")
         getPlayerInfoStmt.setString(1, player.uniqueId.toString())
         return getPlayerInfoStmt.executeQuery()
     }
