--- src/main/java/com/nearvanilla/deathmanager/exceptions/DeathManagerException.kt
@@ -1,4 +1,0 @@
-/* Licensed under GNU General Public License v3.0 */
-package com.nearvanilla.deathmanager.exceptions
-
-class DeathManagerException(message: String) : Exception(message)
